# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Цифровой двойник цеха — точка входа",
    "version": "1.0.0",
    "description": "Единая документированная точка входа двойника. По этим адресам ходят имитаторы 1С в прототипе и пойдут настоящие системы завода.\n\n- Все события проверяются схемой; ошибки сохраняются и видны на экране «Источники данных».\n- Повтор события с тем же `eventId` не создаёт дубль (идемпотентность).\n- Время — ISO 8601 с часовым поясом, завод работает по UTC+5.\n- Контроллеры и видеоаналитика подключаются по MQTT, см. «Документация MQTT (AsyncAPI)».\n\nТипичная интеграция с 1С:MES/QLS/WMS — HTTP-сервисы или OData; конкретная выгрузка настраивается со специалистами завода."
  },
  "servers": [
    {
      "url": "/",
      "description": "Шлюз двойника"
    }
  ],
...
  "paths": {
...
    "/api/v1/quality": {
      "post": {
        "tags": [
          "1С и мастер"
        ],
        "summary": "Итог качества в формате выданной таблицы",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityRequest"
              },
              "examples": {
                "Данные организаторов": {
                  "value": [
                    {
                      "date": "01.10.2026",
                      "area": "Сварка",
                      "produced": 118,
                      "defects": 2,
                      "pct": 1.7
                    },
                    {
                      "date": "01.10.2026",
                      "area": "Окраска",
                      "produced": 115,
                      "defects": 4,
                      "pct": 3.5
                    },
                    {
                      "date": "01.10.2026",
                      "area": "Сборка",
                      "produced": 121,
                      "defects": 1,
                      "pct": 0.8
                    },
                    {
                      "date": "02.10.2026",
                      "area": "Сварка",
                      "produced": 111,
                      "defects": 3,
                      "pct": 2.7
                    },
                    {
                      "date": "02.10.2026",
                      "area": "Окраска",
                      "produced": 116,
                      "defects": 6,
                      "pct": 5.2
                    },
                    {
                      "date": "02.10.2026",
                      "area": "Сборка",
                      "produced": 119,
                      "defects": 2,
                      "pct": 1.7
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Итог приёма и найденные противоречия",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vin/{vin}": {
      "get": {
        "tags": [
          "Двойник"
        ],
        "summary": "Паспорт автомобиля: маршрут по постам и условия в момент прохода",
        "parameters": [
          {
            "name": "vin",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Vin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Паспорт",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "VIN не найден"
          }
        }
      }
    },
...
    "/api/v1/bodies/{id}": {
      "get": {
        "tags": [
          "Двойник"
        ],
        "summary": "Кузов по номеру или VIN: маршрут операций, история отметок, восстановленные отметки и петли перекраски",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Номер кузова (B-04812) или VIN"
          }
        ],
        "responses": {
          "200": {
            "description": "Кузов",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Кузов не найден"
          }
        }
      }
    },
...
    "/api/v1/decisions": {
      "post": {
        "tags": [
          "Двойник"
        ],
        "summary": "Принять вариант решения по инциденту",
        "description": "Двойник выдаёт наряд в системы завода (MQTT allur/kst/twin/work-orders) и пересчитывает прогноз.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              },
              "examples": {
                "Заменить фильтр в пересменку": {
                  "value": {
                    "incidentId": "inc-quality-paint-0001",
                    "optionId": "replace_at_shift_change",
                    "decidedBy": "Начальник производства"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Решение принято, наряд выдан",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка валидации (сообщения по-русски)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Инцидент или вариант не найден"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Vin": {
        "type": "string",
        "pattern": "^[A-HJ-NPR-Z0-9]{17}$",
        "description": "VIN кузова, 17 символов",
        "examples": [
          "KZACN1S18TK004812"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          }
        },
        "required": [
          "error",
          "message"
        ]
      },
...
      "DateLike": {
        "type": "string",
        "pattern": "^(\\d{4}-\\d{2}-\\d{2}|\\d{1,2}\\.\\d{1,2}\\.\\d{4})$"
      },
...
      "QualityRow": {
        "type": "object",
        "properties": {
          "date": {
            "$ref": "#/components/schemas/DateLike"
          },
          "shift": {
            "anyOf": [
              {
                "type": "number",
                "const": 1
              },
              {
                "type": "number",
                "const": 2
              }
            ]
          },
          "area": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "description": "Участок: «Сварка», «Окраска», «Сборка» или код (weld…)"
          },
          "produced": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "defects": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "pct": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          }
        },
        "required": [
          "date",
          "area",
          "produced",
          "defects"
        ],
        "description": "Строка итога качества в формате выданной таблицы"
      },
      "QualityRequest": {
        "minItems": 1,
        "maxItems": 5000,
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/QualityRow"
        }
      },
      "DecisionRequest": {
        "type": "object",
        "properties": {
          "incidentId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "optionId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "decidedBy": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "incidentId",
          "optionId"
        ],
        "description": "Принятие варианта решения по инциденту"
      },
...
    },
...
  }
}